# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface AuthInput {
  email: string;
  /** @minLength 10 */
  password: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 10 */
  password: string;
  /** @minLength 1 */
  displayName: string;
}

export type CircleTier = typeof CircleTier[keyof typeof CircleTier];


export const CircleTier = {
  non_assisted: 'non_assisted',
  transitional: 'transitional',
  fully_assisted: 'fully_assisted',
} as const;

export type CircleRole = typeof CircleRole[keyof typeof CircleRole];


export const CircleRole = {
  primary_user: 'primary_user',
  primary_caretaker: 'primary_caretaker',
  primary_physician: 'primary_physician',
  family: 'family',
} as const;

export interface Circle {
  id: string;
  name: string;
  recipientName: string;
  tier: CircleTier;
  role: CircleRole;
}

export interface SessionUser {
  id: string;
  email: string;
  displayName: string;
  circles: Circle[];
}

export interface CareRecord {
  id: string;
  kind: string;
  title: string;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  dueAt?: string | null;
  createdAt: string;
}

export interface RecordInput {
  /** @minLength 1 */
  kind: string;
  /** @minLength 1 */
  title: string;
  details?: string;
  dueAt?: string;
}

export type ApprovalStatus = typeof ApprovalStatus[keyof typeof ApprovalStatus];


export const ApprovalStatus = {
  pending: 'pending',
  applied: 'applied',
  rejected: 'rejected',
  expired: 'expired',
} as const;

export interface Approval {
  id: string;
  circleId: string;
  action: string;
  status: ApprovalStatus;
}

export type ApprovalInputAction = typeof ApprovalInputAction[keyof typeof ApprovalInputAction];


export const ApprovalInputAction = {
  create_record: 'create_record',
} as const;

export type ApprovalInputPayload = { [key: string]: unknown };

export interface ApprovalInput {
  action: ApprovalInputAction;
  payload: ApprovalInputPayload;
}

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  primary_user: 'primary_user',
  primary_caretaker: 'primary_caretaker',
  primary_physician: 'primary_physician',
  family: 'family',
} as const;

export interface Member {
  id: string;
  userId: string;
  displayName?: string;
  email?: string;
  role: MemberRole;
}

export type MemberInputRole = typeof MemberInputRole[keyof typeof MemberInputRole];


export const MemberInputRole = {
  primary_user: 'primary_user',
  primary_caretaker: 'primary_caretaker',
  primary_physician: 'primary_physician',
  family: 'family',
} as const;

export interface MemberInput {
  userId: string;
  role: MemberInputRole;
}

export interface CaretakerInput {
  userId: string;
}

export type TierInputTier = typeof TierInputTier[keyof typeof TierInputTier];


export const TierInputTier = {
  non_assisted: 'non_assisted',
  transitional: 'transitional',
  fully_assisted: 'fully_assisted',
} as const;

export interface TierInput {
  tier: TierInputTier;
}

export interface Audit {
  id: string;
  action: string;
  /** @nullable */
  target?: string | null;
  outcome: string;
  createdAt: string;
}