# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /circles/{circleId}/records:
    get:
      operationId: listRecords
      tags: [records]
      parameters: [{ name: circleId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: Records, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/CareRecord" } } } } }
        "403": { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    post:
      operationId: createRecord
      tags: [records]
      parameters: [{ name: circleId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecordInput" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/CareRecord" } } } }
        "401": { description: Unauthenticated, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  /circles/{circleId}/audit:
    get:
      operationId: listAudit
      tags: [circles]
      parameters: [{ name: circleId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Audit entries, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Audit" } } } } } }
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties: { error: { type: string } }
...
    CareRecord:
      type: object
      required: [id, kind, title, createdAt]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        title: { type: string }
        details: { type: ["string", "null"] }
        dueAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    RecordInput:
      type: object
      required: [kind, title]
      properties:
        kind: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        details: { type: string }
        dueAt: { type: string, format: date-time }
...
    Audit:
      type: object
      required: [id, action, outcome, createdAt]
      properties:
        id: { type: string, format: uuid }
        action: { type: string }
        target: { type: ["string", "null"] }
        outcome: { type: string }
        createdAt: { type: string, format: date-time }