# lib/api-zod/src/generated/types/careRecord.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareRecord {
  id: string;
  kind: string;
  title: string;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  dueAt?: Date | null;
  createdAt: Date;
}