# lib/api-zod/src/generated/types/recordInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecordInput {
  /** @minLength 1 */
  kind: string;
  /** @minLength 1 */
  title: string;
  details?: string;
  dueAt?: Date;
}